+++ tests/unit/api/userManagement.spec.ts
@@ -1,15 +1,19 @@
 import Student from "@/api/api_models/user_management/Student";
+import User_List from "@/api/api_models/user_management/User_List";
 import UserManagement from "@/api/UserManagement";
 import { Account } from "@/entities/Account";
 import { Role } from "@/entities/Role";
+import { store } from "@/use/store/store";
+import { MutationTypes } from "@/use/store/mutation-types";
+import { getRandomizedUserAndAuthUser } from "../../helper/Users";
 import { readFileSync } from "fs";
 import MachineUserAuthenticationManagement from "../../helper/MachineUserAuthenticationManagement";
-import { getRandomizedUserAndAuthUser } from "../../helper/Users";
 
 var userManagement: UserManagement;
-const pair = getRandomizedUserAndAuthUser(Role.STUDENT) as { student: Student; authUser: Account };
+const pair = getRandomizedUserAndAuthUser(Role.STUDENT) as { governmentId: string; student: Student; authUser: Account };
 const student = pair.student;
 const authUser = pair.authUser;
+const governmentId = pair.governmentId;
 
 const adminAuth = JSON.parse(readFileSync("tests/fixtures/logins/admin.json", "utf-8")) as { username: string; password: string };
 const studentAuth = JSON.parse(readFileSync("tests/fixtures/logins/student.json", "utf-8")) as { username: string; password: string };
@@ -30,7 +34,7 @@ beforeAll(async () => {
 });
 
 test("Create user", async () => {
-    const success = await userManagement.createUser(authUser, student);
+    const success = await userManagement.createUser(governmentId, authUser, student);
     expect(success.returnValue).toBe(true);
     await new Promise((r) => setTimeout(r, 1000));
 });
@@ -39,6 +43,7 @@ test("Get specific user", async () => {
     var result = false;
     const user = await userManagement.getSpecificUser(student.username);
     result = user.returnValue.username == student.username;
+    student.enrollmentIdSecret = user.returnValue.enrollmentIdSecret;
     expect(result).toBe(true);
 });
 
@@ -71,44 +76,50 @@ test("Get own user", async () => {
 });
 
 test("Get all users", async () => {
-    const users = await userManagement.getAllUsers();
+    const users = await userManagement.getUsers();
     let result = true;
-    result = result && users.returnValue.students.length > 0;
+    result = result && (users.returnValue as User_List).students.length > 0;
     expect(result).toBe(true);
 });
 
 test("Get all students", async () => {
-    const users = await userManagement.getAllUsersByRole(Role.STUDENT);
-    let result = users.returnValue.length > 0;
+    const users = await userManagement.getUsers(Role.STUDENT);
+    let result = (users.returnValue as Student[]).length > 0;
     expect(result).toBe(true);
 });
 
 test("Get users by usernames", async () => {
-    const users = await userManagement.getUsers(studentAuth.username, lecturerAuth.username);
+    const users = await userManagement.getUsers(undefined, [studentAuth.username, lecturerAuth.username]);
     let result = Object.values(users.returnValue).flat();
     expect(result).toHaveLength(2);
 });
 
 test("Get lecturers by usernames", async () => {
-    const users = await userManagement.getLecturers(studentAuth.username, lecturerAuth.username);
+    const users = await userManagement.getUsers(Role.LECTURER, [studentAuth.username, lecturerAuth.username]);
     let result = Object.values(users.returnValue).flat();
     expect(result).toHaveLength(1);
 });
 
 test("Get empty list of lecturers by usernames", async () => {
-    const users = await userManagement.getLecturers();
+    const users = await userManagement.getUsers(Role.LECTURER, []);
     let result = Object.values(users.returnValue).flat();
     expect(result).toHaveLength(0);
 });
 
 test("Get students by usernames", async () => {
-    const users = await userManagement.getStudents(studentAuth.username, lecturerAuth.username);
+    const users = await userManagement.getUsers(Role.STUDENT, [studentAuth.username, lecturerAuth.username]);
     let result = Object.values(users.returnValue).flat();
     expect(result).toHaveLength(1);
 });
 
 test("Get admins by usernames", async () => {
-    const users = await userManagement.getAdmins(studentAuth.username, lecturerAuth.username, adminAuth.username);
+    const users = await userManagement.getUsers(Role.ADMIN, [studentAuth.username, lecturerAuth.username, adminAuth.username]);
+    let result = Object.values(users.returnValue).flat();
+    expect(result).toHaveLength(1);
+});
+
+test("Get active student", async () => {
+    const users = await userManagement.getUsers(undefined, [student.username], true);
     let result = Object.values(users.returnValue).flat();
     expect(result).toHaveLength(1);
 });
@@ -155,4 +166,27 @@ test("get dummy profile picture", async () => {
 test("Delete user", async () => {
     const success = await userManagement.deleteUser(student.username);
     expect(success.returnValue).toBe(true);
+});
+
+test("Get active user", async () => {
+    const resp = await userManagement.getUsers(undefined, [student.username], true);
+    let result = Object.values(resp.returnValue).flat();
+    expect(result).toHaveLength(0);
+});
+
+test("Get inactive user", async () => {
+    const resp = await userManagement.getUsers(undefined, [student.username], false);
+    let result = Object.values(resp.returnValue).flat();
+    expect(result).toHaveLength(1);
+});
+
+test("Force delete user", async () => {
+    const success = await userManagement.forceDeleteUser(student.username);
+    expect(success.returnValue).toBe(true);
+});
+
+test("Get user", async () => {
+    const resp = await userManagement.getUsers(undefined, [student.username]);
+    let result = Object.values(resp.returnValue).flat();
+    expect(result).toHaveLength(0);
 });